pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "neobert"
version = "0.0.0"
description = "A state-of-the-art update of BERT"
requires-python = ">= 3.10"
dependencies = [
    "torch==2.5.*",
    "accelerate==1.1.*",
    "deepspeed==0.15.4",
    "xformers==0.0.28.*",
    "transformers==4.46.*",
    "datasets==3.1.*",
    "wandb==0.18.*",
    "hydra-core==1.3.*",
    "numpy==1.26.*",
    "scikit-learn==1.5.*",
    "evaluate==0.4.*",
    "mteb",
    "libaio",
    "tqdm",
    "einops",
]
authors = [
    { name = "Lola Le Breton", email = "lola.lebreton@mila.quebec" },
    { name = "Quentin Fournier", email = "quentin.fournier@mila.quebec" },
]
maintainers = [
    { name = "Lola Le Breton", email = "lola.lebreton@mila.quebec" },
    { name = "Quentin Fournier", email = "quentin.fournier@mila.quebec" },
]
readme = "README.md"
license = { text = "MIT License" }
keywords = []

[project.optional-dependencies]
dev = ["black", "pytest", "pytest-cov", "pytest-clarity", "xeger"]